+++ mod_WolfoItemBuffs/CHANGELOG.md
@@ -1,4 +1,24 @@
 # Changelog
 ```
+v1.0.3
+Fixed Leeching Seed healing twice.
+
+Harpoon Speed now 100% (Down from 125%)
+Config to customize Harpoon Speed.
+(125% is a bit much in this game)
+
+Aegis Barrier Decay now 10% (Up from 0%)
+Config to customize Aegis Barrier Decay Mult.
+(Barrier Decaying is kind of it's thing. Don't think this effect needs to be weak and stack)
+
+Laser Scope 15% Crit (Up from 10%)
+REX Shift, Void Fiend Shift; No longer put you into combat (Red Whip)
+
+v1.0.2
+Reduced Singularity band damage to 150% (+150%).
+Reduced Leech Seed Dot heal to 0.2.
+(The values before were definitely too high and not tested)
+
 v1.0.0: Release
 -With Config
+